Create script to convert and relocate a file
The idea at the moment is that this script will be called from a gradle
task on the appropriate files using an exec task.

Ideally I would do this path substitution in gradle, or queue up a
bunch of files to do them in bash in one go, instead of invoking bash
every time, but that can come later. For the moment, getting this
working should work well.

I can also copy non-markdown files using the gradle copy task, and
hopefully create the directory structure that way as well.