Create needed directories with mkdir in script
It’d be nice to move this to a separate gradle step in the future, as I
can do that once and then not have to repeat the check unnecessarily for
separate files in the same folder.

Alternatively, I could make a separate bash script to check and
duplicate the folder structure, which might be handy.