Invoke pandoc directly from gradle
Bash is no longer needed to calculate the output paths from the input
paths, as I can do it in Java – now the gradle script can directly
invoke pandoc in place of the bash script.

The gradle script is currently entirely an ad-hoc task and is not very
configurable. It also should be able to tell if the build is not out of
date, and avoid needing to rebuild, but if it does need to rebuild, it
will rebuild all the files, instead of just the files that need
rebuilding.

I’ll see if incremental build is possible with ad-hoc tasks, but I
suspect the easiest direction to go in now is to convert the ad-hoc task
to a proper task class.